Created the Item Crate Mechanic
I made a new folder in the function section for mechanics and also added my itemCrate mcfunction file to the tick.json file. I also put 2 structure files into the worlds/contributions folder. I added a new scoreboard into the scoreboard mcfunction file and lastly I added another mcfunction which is named itemCrateStarted which when ran will grant the player the starter kit for the itemCrate and a little guide book that can be updated by anyone and has a last updated and by who section.