Add custom frying pan functions to trash bin
Now, when the trash bin is interacted with the frying pan, information about what the frying pan is holding is cleared, effectively clearing the food on it. The Miro board specifies that the trash bin clears burnt food, but currently there is no way to tell if an item is considered "burnt" or not. Functionality may need to change in the future for this reason.