Add pot and base custom component classes
Base custom component classes make it easier to write duplicate behavior across components. NOTE: to use the this keyword in custom components in classes, the methods MUST be defined as an arrow function. If defined as a normal instance method, the this keyword will be null. This seems to be due to how the Script API handles custom components internally.